backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning
tensorflow>=2.16.0
numpy>=1.26.0
pandas==2.1.4
scikit-learn==1.3.2
joblib>=1.3.0

# HTTP requests for API calls
httpx==0.25.2
requests==2.31.0
openmeteo-requests==1.7.2

# Data processing
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0
requests-cache==1.2.1
retry-requests==2.0.0

# Additional utilities
pydantic==2.5.0
typing-extensions==4.8.0